# swagger/openapi.yaml
openapi: 3.0.1
info:
  version: 0.0.1
  title: School API
  description: School API
servers:
  - url: 'http://localhost:3000'
paths:
...
  /subjects:
    get:
      summary: returns subjects collection
      responses:
        '200':
          description: displays subjects collection successfully
        '404':
          description: subjects collection was not found
        '500': 
          description: server not found
    
    post:
      summary: inserting changes and data
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        '200':
          description: changes were implemented successfully 
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    put:
      summary: editing data
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        '200':
          description: changes are saved
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
    delete:
      summary: data was removed
      responses:
        '200':
          description: data was removed successfully
        '404':
          description: data not found
  /subjects/{subjectId}:
    get:
      summary: returns subjectId Id
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: subject Id was successfully returned
        '404':
          description: subjectId is not existed
        '500':
          description: servers not found
          content:
            application/json:
              schema:
                type: object
          
    post:
      summary: input data and changes
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: returns the success status if data insertion or changes were occurred
        '404':
          description: subjectId was not found
        '500':
          description: servers not found
          content:
            application/json:
              schema:
                type: object
          
    put:
      summary: applies modifications and changes
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: displays success status when data were modified
        '403':
          description: editing data is forbidden
        '405':
          description: the editing method is not allowed
        '500':
          description: servers not found
          content:
            application/json:
              schema:
                type: object
          
    delete:
      summary: returns something
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: data was successfully removed
        '500':
          description: servers not found
          content:
            application/json:
              schema:
                type: object
  /subjects/{subjectId}/seasons:
    get:
      summary: returns seasons collection
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: displays seasons collection successfully
        '404':
          description: teachers collection was not found
        '500': 
          description: server not found
    post:
      summary: inserting changes and data
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        '200':
          description: changes were implemented successfully 
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    put:
      summary: editing data
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        '200':
          description: changes are saved
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
    delete:
      summary: data was removed
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: data was removed successfully
        '404':
          description: data not found
  /subjects/{subjectId}/seasons/{seasonId}:
    get:
      summary: returns seasonId Id
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: season Id was successfully returned
        '404':
          description: season Id is not existed
        '500':
          description: server was not found
          content:
            application/json:
              schema:
                type: object
          
    post:
      summary: input data and changes
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '201':
          description: returns the success status if data insertion or changes were occurred
        '404':
          description: subjectId was not found
        '500':
          description: servers not found
          content:
            application/json:
              schema:
                type: object
          
    put:
      summary: applies modifications and changes
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: displays success status when data were modified
        '403':
          description: editing data is forbidden
        '405':
          description: the editing method is not allowed
        '500':
          description: servers not found
          content:
            application/json:
              schema:
                type: object
          
    delete:
      summary: removes season Id
      #need to add the below parameters to the rest of the paths
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: data was successfully removed
        '500':
          description: servers not found
          content:
            application/json:
              schema:
                type: object
  /subjects/{subjectId}/seasons/{seasonId}/lessons:
    get:
      summary: returns lessons collection
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: displays lessons collection successfully
        '404':
          description: lessons collection was not found
        '500': 
          description: server not found
    
    post:
      summary: inserting changes and data
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        '200':
          description: changes were implemented successfully 
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    put:
      summary: editing data
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        '200':
          description: changes are saved
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
    delete:
      summary: data was removed
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: data was removed successfully
        '404':
          description: data not found
  /subjects/{subjectId}/seasons/{seasonId}/lessons/{lessonId}:
    get:
      summary: returns lessonId Id
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: season Id was successfully returned
        '404':
          description: season Id is not existed
        '500':
          description: server was not found
          content:
            application/json:
              schema:
                type: object
          
    post:
      summary: input data and changes
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: returns the success status if data insertion or changes were occurred
        '404':
          description: subjectId was not found
        '500':
          description: servers not found
          content:
            application/json:
              schema:
                type: object
          
    put:
      summary: applies modifications and changes
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: displays success status when data were modified
        '403':
          description: editing data is forbidden
        '405':
          description: the editing method is not allowed
        '500':
          description: servers not found
          content:
            application/json:
              schema:
                type: object
          
    delete:
      summary: removes season Id
      parameters: 
        - name: subjectId
          in: path
          required: true
          schema: 
            type: string
        - name: seasonId
          in: path
          required: true
          schema: 
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: data was successfully removed
        '500':
          description: servers not found
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
    Teacher:
      type: object
      required:
        - hash
      properties: 
        hash: 
          type: string
          minLength: 3
          maxLength: 15
          example: "teacher"
        name:
          type: object
          properties:
            first: 
              type: string
              example: Andrey
            last:
              type: string
              example: Prisniak
        image:
          type: string
          example: http://example.com/image.jpg
        dateOfBirth:
          type: string
          format: date
          example: 22/04/1990
        emails:
          type: array
          items: 
            type: object
            properties:
              email:
                type: string
                example: test@mail.com
              primary:
                type: boolean
                example: true
        phones: 
          type: array
          items:
            type: object
            properties: 
              phone: 
                type: string
                example: to be determined
              primary:
                type: boolean
                example: true
                
        sex: 
          type: string
          enum: [male, female]
          example: female
          
        social:
          type: object
          properties:
            facebook: 
              type: string
              example: aws shkara
            linkedIn: 
              type: string
              example: aws-shkara
            skype: 
              type: string
              example: aws.shkara
            telegram: 
              type: string
              example: Atawsy  
          
        subjects:
          type: array
          items:
            type: object
            properties:
              subject:
                type: string
                example: http://example.com/api/subjects/123
                
        description:
          type: string
          example: information about teachers
          
        started:
          type: string
          format: date
          example: 17/09/2018
          
        created:
          type: string
          format: date
          example: 17/09/2018
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic